+++ CHANGELOG.md
@@ -1,9 +1,28 @@
 # CHANGELOG
 
-Pandashark's Portfolio の大まかな機能変化記録
+Pandashark's Portfolioの大まかな機能変化記録
+
+## 2022-08-06
+
+- スキル更新
+
+## 2022-07-14
+
+- google-form-notificationを追加
+- 個人開発ページのグルーピングを再構成
+
+## 2022-05-04
+
+- Cryptocurrency Trackerを追加
+
+## 2022-04-23
+
+### Product Added
+
+- Simple Kanban Plannerを追加
 
 ## 2022-03-13
 
 ### Development Added
 
-- サイト作成して Netlify で公開
+- サイト作成してNetlifyで公開